% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{beta_binomial_estimator}
\alias{beta_binomial_estimator}
\title{Beta-Binomial Chiu estimator}
\usage{
beta_binomial_estimator(data_mat)
}
\arguments{
\item{obs_n}{\link{list} number of new observed}

\item{est_n}{\link{list} number of estimated new}

\item{obs_t}{\link{numeric} number of observed in training set}
}
\description{
Beta-Binomial Chiu estimator
}
