% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{buffet_gamma_IBP}
\alias{buffet_gamma_IBP}
\title{Buffet procedure for IBP with Gamma mixture from beginning}
\usage{
buffet_gamma_IBP(alpha, theta, n, a, b)
}
\arguments{
\item{alpha}{value of alpha in product-form feature allocation}

\item{theta}{value of theta in product-form feature allocation}

\item{n}{dimension of the sample to simulate}

\item{a}{Gamma hyperparameter (shape)}

\item{b}{Gamma hyperparameter (rate)}
}
\value{
list: $features contains the simulated features for each customer,
$num_new contains the number of new features selected for each customer
$counts contains the counts for the observed features
}
\description{
Buffet procedure for IBP with Gamma mixture from beginning
}
