% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{buffet_gamma_IBP_initial_sample}
\alias{buffet_gamma_IBP_initial_sample}
\title{Buffet procedure for IBP with Gamma mixture given initial sample}
\usage{
buffet_gamma_IBP_initial_sample(alpha, theta, m, n, counts, a, b)
}
\arguments{
\item{alpha}{value of alpha in product-form feature allocation}

\item{theta}{value of theta in product-form feature allocation}

\item{m}{dimension of the new sample to be observed}

\item{n}{dimension of the already observed sample}

\item{counts}{vector of counts for the already observed features}

\item{a}{Gamma hyperparameter (shape)}

\item{b}{Gamma hyperparameter (rate)}
}
\value{
list: $features contains the simulated features for each customer,
$num_new contains the number of new features selected for each customer
}
\description{
Buffet procedure for IBP with Gamma mixture given initial sample
}
