% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{buffet_negbin_BB}
\alias{buffet_negbin_BB}
\title{Buffet procedure for BB with Negative-Binomial mixture from beginning}
\usage{
buffet_negbin_BB(alpha, theta, n, nstar, p)
}
\arguments{
\item{alpha}{value of alpha in product-form feature allocation}

\item{theta}{value of theta in product-form feature allocation}

\item{n}{dimension of the sample to simulate}

\item{nstar}{Negative-Binomial hyperparameter (number of successes)}

\item{p}{Negative-Binomial hyperparameter (success probability)}
}
\value{
list: $features contains the simulated features for each customer,
$num_new contains the number of new features selected for each customer,
$counts contains the counts for the observed features
}
\description{
Buffet procedure for BB with Negative-Binomial mixture from beginning
}
