% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{buffet_negbin_BB_initial_sample}
\alias{buffet_negbin_BB_initial_sample}
\title{Buffet procedure for BB with Negative-Binomial mixture given initial sample}
\usage{
buffet_negbin_BB_initial_sample(alpha, theta, m, n, counts, nstar, p)
}
\arguments{
\item{alpha}{value of alpha in product-form feature allocation}

\item{theta}{value of theta in product-form feature allocation}

\item{m}{dimension of the new sample to be observed}

\item{n}{dimension of the already observed sample}

\item{counts}{vector of counts for the already observed features}

\item{nstar}{Negative-Binomial hyperparameter (number of successes)}

\item{p}{Negative-Binomial hyperparameter (success probability)}
}
\value{
list: $features contains the simulated features for each customer,
$num_new contains the number of new features selected for each customer
$counts contains the counts for the observed features
}
\description{
Buffet procedure for BB with Negative-Binomial mixture given initial sample
}
