% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{buffet_poiss_BB}
\alias{buffet_poiss_BB}
\title{Buffet procedure for BB with Poisson mixture from beginning}
\usage{
buffet_poiss_BB(alpha, theta, n, lambda)
}
\arguments{
\item{alpha}{value of alpha in product-form feature allocation}

\item{theta}{value of theta in product-form feature allocation}

\item{n}{dimension of the sample to simulate}

\item{lambda}{Poisson hyperparameter}
}
\value{
list: $features contains the simulated features for each customer,
$num_new contains the number of new features selected for each customer
$counts contains the counts for the observed features
}
\description{
Buffet procedure for BB with Poisson mixture from beginning
}
