% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines_gamma_IBP.R
\name{CI_Kmn_gamma_IBP}
\alias{CI_Kmn_gamma_IBP}
\title{Credible intervals of Kmn for the IBP with Gamma mixture}
\usage{
CI_Kmn_gamma_IBP(alpha, theta, m, n, Kn, a, b, lev)
}
\arguments{
\item{alpha}{\link{numeric} value of alpha in product-form feature allocation}

\item{theta}{\link{numeric} value of theta in product-form feature allocation}

\item{m}{\link{numeric} dimension of the new sample to be observed}

\item{n}{\link{numeric} dimension of the already observed sample}

\item{Kn}{\link{numeric} number of features in the already observed sample}

\item{a}{\link{numeric} Gamma hyperparameter (shape)}

\item{b}{\link{numeric} Gamma hyperparameter (rate)}

\item{lev}{\link{numeric} level of the credible intervals}
}
\description{
This function computes the means and the credible intervals of Kmn for the IBP with Gamma mixture
}
