% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines_negbin_BB.R
\name{CI_Kmn_negbin_BB}
\alias{CI_Kmn_negbin_BB}
\title{Credible intervals of Kmn for the BB with Negative-Binomial mixture}
\usage{
CI_Kmn_negbin_BB(alpha, theta, m, n, Kn, nstar, p, lev)
}
\arguments{
\item{alpha}{\link{numeric} value of alpha in product-form feature allocation}

\item{theta}{\link{numeric} value of theta in product-form feature allocation}

\item{m}{\link{numeric} dimension of the new sample to be observed}

\item{n}{\link{numeric} dimension of the already observed sample}

\item{Kn}{\link{numeric} number of features in the already observed sample}

\item{nstar}{\link{numeric} Negative-Binomial hyperparameter (number of successes)}

\item{p}{\link{numeric} Negative-Binomial hyperparameter (success probability)}

\item{lev}{\link{numeric} level of the credible intervals}
}
\description{
This function computes the means and the credible intervals of Kmn for the BB with Negative-Binomial mixture
}
