% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines_poiss_BB.R
\name{CI_Kmn_poiss_BB}
\alias{CI_Kmn_poiss_BB}
\title{Credible intervals of Kmn for the BB with Poisson mixture}
\usage{
CI_Kmn_poiss_BB(alpha, theta, m, n, lambda, lev)
}
\arguments{
\item{alpha}{\link{numeric} value of alpha in product-form feature allocation}

\item{theta}{\link{numeric} value of theta in product-form feature allocation}

\item{m}{\link{numeric} dimension of the new sample to be observed}

\item{n}{\link{numeric} dimension of the already observed sample}

\item{lambda}{\link{numeric} Poisson hyperparameter}

\item{lev}{\link{numeric} level of the credible intervals}
}
\description{
This function computes the means and the credible intervals of Kmn for the BB with Poisson mixture
}
