% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_accuracy}
\alias{compute_accuracy}
\title{Metric of accuracy in prediction}
\usage{
compute_accuracy(obs_n, est_n, obs_t)
}
\arguments{
\item{obs_n}{\link{list} number of new observed}

\item{est_n}{\link{list} number of estimated new}

\item{obs_t}{\link{numeric} number of observed in training set}
}
\description{
Metric of accuracy in prediction
}
