% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_choice.R
\name{compute_BF}
\alias{compute_BF}
\title{Compute Bayes Factor (BF) for NegBinBB/GammaIBP (Fully-Bayesian approach)}
\usage{
compute_BF()
}
\arguments{
\item{model_0_fit}{object of class \code{GibbsFA, NegBinBB} or \code{GibbsFA, GammaIBP}}

\item{model_1_fit}{object of class \code{GibbsFA, NegBinBB} or \code{GibbsFA, GammaIBP}}
}
\description{
Compute Bayes Factor (BF) for NegBinBB/GammaIBP (Fully-Bayesian approach)
}
