% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines_gamma_IBP_gibbs.R
\name{compute_der_log_full_sb_sp}
\alias{compute_der_log_full_sb_sp}
\title{Function the gradient of the log full-conditional for s_hat and alpha_bar_hat}
\usage{
compute_der_log_full_sb_sp(
  alpha_hat,
  c,
  beta,
  Gam,
  n,
  K,
  counts,
  a_alpha,
  b_alpha
)
}
\arguments{
\item{n}{}

\item{K}{}

\item{counts}{}

\item{a_alpha}{}

\item{b_alpha}{}
}
\description{
Function the gradient of the log full-conditional for s_hat and alpha_bar_hat
}
