% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines_gamma_IBP_gibbs.R
\name{compute_grad_log_full_gamma_ibp}
\alias{compute_grad_log_full_gamma_ibp}
\title{Function the gradient of the log full-conditional for s_hat and alpha_bar_hat}
\usage{
compute_grad_log_full_gamma_ibp(
  s_hat,
  alpha_hat,
  Gam,
  n,
  K,
  counts,
  a_s,
  b_s,
  a_alpha,
  b_alpha
)
}
\arguments{
\item{s_hat}{}

\item{n}{}

\item{K}{}

\item{counts}{}

\item{a_s}{}

\item{b_s}{}

\item{a_alpha}{}

\item{b_alpha}{}

\item{alpha_bar_hat}{}

\item{lambda}{}
}
\description{
Function the gradient of the log full-conditional for s_hat and alpha_bar_hat
}
