% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines_negbin_BB_gibbs.R
\name{compute_grad_log_full_negbin}
\alias{compute_grad_log_full_negbin}
\title{Function the gradient of the log full-conditional for s_hat and alpha_bar_hat}
\usage{
compute_grad_log_full_negbin(
  nstar_hat,
  s_hat,
  alpha_bar_hat,
  p,
  n,
  K,
  counts,
  a_star,
  b_star,
  a_s,
  b_s,
  a_alpha,
  b_alpha
)
}
\arguments{
\item{s_hat}{}

\item{alpha_bar_hat}{}

\item{n}{}

\item{K}{}

\item{counts}{}

\item{a_s}{}

\item{b_s}{}

\item{a_alpha}{}

\item{b_alpha}{}

\item{lambda}{}
}
\description{
Function the gradient of the log full-conditional for s_hat and alpha_bar_hat
}
