% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines_gamma_IBP_gibbs.R
\name{compute_log_ratio_full_gamma_ibp}
\alias{compute_log_ratio_full_gamma_ibp}
\title{Function computing the log ratio of the full-conditional of s_hat, alpha_bar_hat
in proposed and current values}
\usage{
compute_log_ratio_full_gamma_ibp(
  s_hat_prop,
  alpha_hat_prop,
  s_hat_curr,
  alpha_hat_curr,
  Gam,
  n,
  K,
  counts,
  a_s,
  b_s,
  a_alpha,
  b_alpha
)
}
\arguments{
\item{s_hat_prop}{}

\item{alpha_hat_prop}{}

\item{s_hat_curr}{}

\item{alpha_hat_curr}{}

\item{Gam}{}

\item{n}{}

\item{K}{}

\item{counts}{}

\item{a_s}{}

\item{b_s}{}

\item{a_alpha}{}

\item{b_alpha}{}
}
\description{
Function computing the log ratio of the full-conditional of s_hat, alpha_bar_hat
in proposed and current values
}
