% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines_negbin_BB_gibbs.R
\name{compute_log_ratio_full_negbin}
\alias{compute_log_ratio_full_negbin}
\title{Function computing the log ratio of the full-conditional of s_hat, alpha_bar_hat
in proposed and current values}
\usage{
compute_log_ratio_full_negbin(
  nstar_hat_prop,
  s_hat_prop,
  alpha_bar_hat_prop,
  nstar_hat_curr,
  s_hat_curr,
  alpha_bar_hat_curr,
  p,
  n,
  K,
  counts,
  a_star,
  b_star,
  a_s,
  b_s,
  a_alpha,
  b_alpha
)
}
\arguments{
\item{s_hat_prop}{}

\item{alpha_bar_hat_prop}{}

\item{s_hat_curr}{}

\item{alpha_bar_hat_curr}{}

\item{n}{}

\item{K}{}

\item{counts}{}

\item{a_s}{}

\item{b_s}{}

\item{a_alpha}{}

\item{b_alpha}{}

\item{lambda}{}
}
\description{
Function computing the log ratio of the full-conditional of s_hat, alpha_bar_hat
in proposed and current values
}
