% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines_gamma_IBP_gibbs.R
\name{compute_log_ratio_full_sb_sp}
\alias{compute_log_ratio_full_sb_sp}
\title{Function computing the log ratio of the full-conditional of s_hat, alpha_bar_hat
in proposed and current values}
\usage{
compute_log_ratio_full_sb_sp(
  alpha_hat_prop,
  alpha_hat_curr,
  c,
  beta,
  Gam,
  n,
  K,
  counts,
  a_alpha,
  b_alpha
)
}
\arguments{
\item{Gam}{}

\item{n}{}

\item{K}{}

\item{counts}{}

\item{a_alpha}{}

\item{b_alpha}{}
}
\description{
Function computing the log ratio of the full-conditional of s_hat, alpha_bar_hat
in proposed and current values
}
