% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines_gamma_IBP_gibbs.R
\name{compute_log_ratio_q_gamma_ibp}
\alias{compute_log_ratio_q_gamma_ibp}
\title{Function computing the log of the ratio of the terms related to q in MALA}
\usage{
compute_log_ratio_q_gamma_ibp(
  s_hat_prop,
  alpha_hat_prop,
  s_hat_curr,
  alpha_hat_curr,
  tau,
  grad_log_full_curr,
  grad_log_full_prop
)
}
\arguments{
\item{s_hat_prop}{}

\item{alpha_hat_prop}{}

\item{s_hat_curr}{}

\item{alpha_hat_curr}{}

\item{tau}{}

\item{grad_log_full_curr}{}

\item{grad_log_full_prop}{}
}
\description{
Function computing the log of the ratio of the terms related to q in MALA
}
