% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines_negbin_BB_gibbs.R
\name{compute_log_ratio_q_negbin_precond}
\alias{compute_log_ratio_q_negbin_precond}
\title{Function computing the log of the ratio of the terms related to q in MALA}
\usage{
compute_log_ratio_q_negbin_precond(
  params_prop,
  params_curr,
  tau,
  cov_post,
  grad_log_full_curr,
  grad_log_full_prop
)
}
\arguments{
\item{tau}{}

\item{grad_log_full_curr}{}

\item{grad_log_full_prop}{}

\item{s_hat_prop}{}

\item{alpha_bar_hat_prop}{}

\item{s_hat_curr}{}

\item{alpha_bar_hat_curr}{}
}
\description{
Function computing the log of the ratio of the terms related to q in MALA
}
