% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertfeatures.R
\name{convert_features_matrix}
\alias{convert_features_matrix}
\title{Create features matrix}
\usage{
convert_features_matrix(feature_list)
}
\arguments{
\item{feature_list}{\link{list} list containing the observed features}
}
\description{
This function allows to create the matrix of the
observed features in order-of-appearance, with 1 if present, 0 if not
}
