% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_features_list}
\alias{create_features_list}
\title{Convert matrix of 0/1 to list of individuals with observed features}
\usage{
create_features_list(mat)
}
\arguments{
\item{mat}{\link{integer} matrix of 0/1 to be converted}
}
\description{
This function allows to create the list of individuals with observed features
from the matrix of 0/1
}
