% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_features_matrix.R
\name{create_features_matrix}
\alias{create_features_matrix}
\title{Create features matrix}
\usage{
create_features_matrix(feat_list)
}
\arguments{
\item{feat_list}{\link{list} feat_list$features contains the observed features,
feat_list$num_new contains the number of new features for each customer,
feat_list$counts contains the counts for the observed features}
}
\description{
This function allows to create the matrix of the
observed features in order-of-appearance, with 1 if present, 0 if not
}
