% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines_negbin_BB.R
\name{EB_EFPF_fixed_negbin_mv_BB}
\alias{EB_EFPF_fixed_negbin_mv_BB}
\title{EB based on EFPF-max (possibly some parameters are fixed) -
\itemize{
\item BB with Negative-Binomial mixture (mean-variance parametrization)
}}
\usage{
EB_EFPF_fixed_negbin_mv_BB(n, counts, pars_0, opt = c(T, T, T, T))
}
\arguments{
\item{n}{\link{integer} dimension of the observed sample}

\item{counts}{\link{numeric} vector of cardinalities for the observed features}

\item{pars_0}{\link{numeric} Initialization for (alpha, theta, mu, sigma2) to
be optimized}

\item{opt}{\link{logical} vector with 1 if value is optimized, 0 if not optimized}
}
\description{
This function returns the value of the parameters maximizing the
EFPF for the given sample (possibly some parameters are fixed) -
\itemize{
\item BB with Negative-Binomial mixture (mean-variance parametrization)
}
}
