% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines_gamma_IBP.R
\name{EB_EFPF_gamma_IBP}
\alias{EB_EFPF_gamma_IBP}
\title{EB based on EFPF-max - IBP with Gamma mixture}
\usage{
EB_EFPF_gamma_IBP(n, counts, pars_0)
}
\arguments{
\item{n}{\link{integer} dimension of the observed sample}

\item{counts}{\link{numeric} vector of cardinalities for the observed features}

\item{pars_0}{\link{numeric} Initialization for (alpha, theta, a, b) to
be optimized}
}
\description{
This function returns the value of (alpha, theta, a, b) maximizing the
EFPF for the given sample - IBP with Gamma mixture
}
