% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines_negbin_BB.R
\name{EB_EFPF_negbin_BB}
\alias{EB_EFPF_negbin_BB}
\title{EB based on EFPF-max - BB with Negative-Binomial mixture}
\usage{
EB_EFPF_negbin_BB(n, counts, pars_0)
}
\arguments{
\item{n}{\link{integer} dimension of the observed sample}

\item{counts}{\link{numeric} vector of cardinalities for the observed features}

\item{pars_0}{\link{numeric} Initialization for (alpha, theta, n*, p) to
be optimized}
}
\description{
This function returns the value of (alpha, theta, n*, p) maximizing the
EFPF for the given sample - BB with Negative-Binomial mixture
}
