% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines_poiss_BB.R
\name{EB_EFPF_poiss_BB}
\alias{EB_EFPF_poiss_BB}
\title{EB based on EFPF-max - BB with Poisson mixture}
\usage{
EB_EFPF_poiss_BB(n, counts, pars_0)
}
\arguments{
\item{n}{\link{integer} dimension of the observed sample}

\item{counts}{\link{numeric} vector of cardinalities for the observed features}

\item{pars_0}{\link{numeric} Initialization for (alpha, theta, lambda) to
be optimized}
}
\description{
This function returns the value of (alpha, theta, lambda) maximizing the
EFPF for the given sample - BB with Poisson mixture
}
