% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines_poiss_BB.R
\name{EB_MM_cv_poiss_BB}
\alias{EB_MM_cv_poiss_BB}
\title{EB based on method of moments with cross-validation - BB with Poisson mixture}
\usage{
EB_MM_cv_poiss_BB(n, nfold, obs_sample, pars_0, seed_id)
}
\arguments{
\item{n}{\link{integer} dimension of the observed sample}

\item{nfold}{\link{integer} number of folds}

\item{obs_sample}{\link{list} list with the whole sample of features, where
$features contains the simulated features for each customer,
$num_new contains the number of new features selected for each customer
$counts contains the counts for the observed features}

\item{pars_0}{\link{numeric} Initialization for (alpha, theta, lambda) to
be optimized}

\item{seed_id}{\link{numeric} Seed to be used}
}
\description{
This function returns the value of (alpha, theta, lambda) obtained
via the method of moments with cross-validation for the given sample -
BB with Poisson mixture
}
