% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines_negbin_BB.R
\name{EB_MM_negbin_mv_BB}
\alias{EB_MM_negbin_mv_BB}
\title{EB based on method of moments - BB with Negative-Binomial mixture}
\usage{
EB_MM_negbin_mv_BB(n, ntrain, val_rep, data_list, pars_0, seed_id)
}
\arguments{
\item{n}{\link{integer} dimension of the observed sample}

\item{ntrain}{\link{integer} dimension of training set}

\item{val_rep}{\link{integer} number of samples in the validation averaged over for
assessing the number of new features observed}

\item{data_list}{\link{list} list of features by individual}

\item{pars_0}{\link{numeric} Initialization for (alpha, theta, mu, sigma2) to
be optimized}

\item{seed_id}{\link{numeric} Seed to be used}
}
\description{
This function returns the value of (alpha, theta, mu, sigma2) obtained
via the method of moments for the given sample - BB with NB mixture
}
