% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrapolation.R
\name{extrapolation.GammaIBP}
\alias{extrapolation.GammaIBP}
\title{Model-based extrapolation for IBP with Gamma(a,b) mixture,
with prior on a and b}
\usage{
\method{extrapolation}{GammaIBP}(object, M = 10, only_last = FALSE, seed = 1234)
}
\arguments{
\item{object}{An object of class \code{GibbsFA, GammaIBP}}

\item{M}{Maximum size of future sample to extrapolate}

\item{only_last}{logical; TRUE if only extrapolation for the additional sample of size M}

\item{seed}{Set seed for randomness}
}
\description{
Model-based extrapolation for IBP with Gamma(a,b) mixture,
with prior on a and b
}
\details{
Draw samples from the posterior distribution of k_n + K^(n)_m, for different m,
for IBP with Gamma(a,b) mixture, and prior on a and b
}
