% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrapolation.R
\name{extrapolation.GibbsFA}
\alias{extrapolation.GibbsFA}
\title{Model-based extrapolation for GibbsFA models}
\usage{
\method{extrapolation}{GibbsFA}(object, M = 10, only_last = FALSE, ...)
}
\arguments{
\item{object}{An object of class \code{GibbsFA}}

\item{M}{Maximum size of future sample to extrapolate}

\item{only_last}{logical; TRUE if only extrapolation for the additional sample of size M}

\item{seed}{Set seed for randomness}
}
\description{
Model-based extrapolation for GibbsFA models
}
\details{
Draw samples from the posterior distribution of k_n + K^(n)_m, for different m,
for the GibbsFA models
}
