% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrapolation.R
\name{extrapolation.NegBinBB}
\alias{extrapolation.NegBinBB}
\title{Model-based extrapolation for BB with NB(n0,mu0) mixture}
\usage{
\method{extrapolation}{NegBinBB}(object, M = 10, seed = 1234)
}
\arguments{
\item{object}{An object of class \code{GibbsFA, NegBinBB}}

\item{M}{Maximum size of future sample to extrapolate}

\item{seed}{Set seed for randomness}
}
\description{
Model-based extrapolation for BB with NB(n0,mu0) mixture
}
\details{
Draw samples from the posterior distribution of k_n + K^(n)_m, for different m,
for BB with NB(n0,mu0) mixture
}
