% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrapolation.R
\name{extrapolation}
\alias{extrapolation}
\title{Model-based extrapolation}
\usage{
extrapolation(object, ...)
}
\arguments{
\item{object}{An object of class \code{GibbsFA}}

\item{...}{Additional parameters}
}
\description{
Model-based extrapolation
}
