% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines_gamma_IBP_gibbs.R
\name{generate_coverage_prob_chain_gamma_ibp}
\alias{generate_coverage_prob_chain_gamma_ibp}
\title{Generate the chain for coverage probability, given the output chains of the mcmc}
\usage{
generate_coverage_prob_chain_gamma_ibp(
  a_chain,
  b_chain,
  alpha_chain,
  theta_chain,
  n,
  Kn = 0
)
}
\arguments{
\item{Kn}{}
}
\description{
Generate the chain for coverage probability, given the output chains of the mcmc
}
