% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines_negbin_BB_gibbs.R
\name{generate_Kmn_chain_negbin}
\alias{generate_Kmn_chain_negbin}
\title{Generate the chain for Kmn, from m=1 to m=M, given the output chains of the mcmc}
\usage{
generate_Kmn_chain_negbin(nstar, p, alpha_chain, theta_chain, M, n, Kn = 0)
}
\arguments{
\item{Kn}{}
}
\description{
Generate the chain for Kmn, from m=1 to m=M, given the output chains of the mcmc
}
