% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines_negbin_BB_gibbs.R
\name{generate_Kmn_chain_negbin_last}
\alias{generate_Kmn_chain_negbin_last}
\title{Generate the chain for Kmn, for m=M, given the output chains of the mcmc}
\usage{
generate_Kmn_chain_negbin_last(
  nstar,
  p,
  alpha_chain,
  theta_chain,
  M,
  n,
  Kn = 0
)
}
\arguments{
\item{Kn}{}
}
\description{
Generate the chain for Kmn, for m=M, given the output chains of the mcmc
}
