% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines_poiss_BB_gibbs.R
\name{generate_Kmn_chain_poiss}
\alias{generate_Kmn_chain_poiss}
\title{Generate the chain for Kmn, from m=1 to m=M, given the output chains of the mcmc}
\usage{
generate_Kmn_chain_poiss(lambda_chain, alpha_chain, theta_chain, M, n)
}
\arguments{
\item{n}{}
}
\description{
Generate the chain for Kmn, from m=1 to m=M, given the output chains of the mcmc
}
