% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines_poiss_BB_gibbs.R
\name{generate_Kmn_chain_poiss_last}
\alias{generate_Kmn_chain_poiss_last}
\title{Generate the chain for Kmn, for m=M, given the output chains of the mcmc}
\usage{
generate_Kmn_chain_poiss_last(lambda, alpha_chain, theta_chain, M, n)
}
\arguments{
\item{n}{}
}
\description{
Generate the chain for Kmn, for m=M, given the output chains of the mcmc
}
