% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines_negbin_BB_gibbs.R
\name{generate_Ntilde_chain_negbin}
\alias{generate_Ntilde_chain_negbin}
\title{Title}
\usage{
generate_Ntilde_chain_negbin(
  nstar_chain_negbin,
  p_chain_negbin,
  alpha_chain_negbin,
  theta_chain_negbin,
  n,
  Kn
)
}
\arguments{
\item{n}{}

\item{Kn}{}

\item{lambda_chain_poiss}{}

\item{alpha_chain_poiss}{}

\item{theta_chain_poiss}{}
}
\description{
Title
}
