% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines_gamma_IBP_gibbs.R
\name{gibbs_sampler_gamma_ibp}
\alias{gibbs_sampler_gamma_ibp}
\title{Metropolis-within-Gibbs sampler for BB + NegBin (Geometric prior on nstar)}
\usage{
gibbs_sampler_gamma_ibp(
  Z,
  a_0,
  b_0,
  s_0,
  alpha_0,
  p,
  r,
  t,
  a_alpha,
  b_alpha,
  a_s,
  b_s,
  tau,
  fixed,
  S,
  n_burnin,
  thin,
  seed
)
}
\arguments{
\item{Z}{\link{integer} binary matrix of presence/absence (n x K- dimensional)}

\item{a_0}{\link{numeric} initial value of a}

\item{b_0}{\link{numeric} initial value of b}

\item{s_0}{\link{numeric} initial value of s}

\item{alpha_0}{\link{numeric} initial value of alpha}

\item{p}{\link{numeric} hyperparameter of Geometric prior on a}

\item{r}{\link{numeric}}

\item{t}{\link{numeric}}

\item{a_alpha}{\link{numeric}}

\item{b_alpha}{\link{numeric}}

\item{a_s}{\link{numeric}}

\item{b_s}{\link{numeric}}

\item{tau}{\link{numeric} MALA step-size}

\item{fixed}{\link{logical} vector indicating if the variable is fixed or has prior:
if fixed, it stays equal to initial value - order: a, b, alpha, s}

\item{S}{\link{integer} number of iterations for the MCMC algorithm}

\item{n_burnin}{\link{integer} number of iterations for the burn-in}

\item{thin}{\link{integer} thinning}

\item{seed}{\link{integer} seed}
}
\description{
Metropolis-within-Gibbs sampler for BB + NegBin (Geometric prior on nstar)
}
