% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines_negbin_BB_gibbs.R
\name{gibbs_sampler_negbin_gamma}
\alias{gibbs_sampler_negbin_gamma}
\title{Metropolis-within-Gibbs sampler for BB + NegBin (Gamma prior on nstar)}
\usage{
gibbs_sampler_negbin_gamma(
  Z,
  nstar_0,
  p_0,
  s_0,
  alpha_bar_0,
  a_star,
  b_star,
  alpha_p,
  beta_p,
  a_alpha,
  b_alpha,
  a_s,
  b_s,
  tau,
  fixed,
  S,
  n_burnin,
  thin,
  seed
)
}
\arguments{
\item{Z}{\link{integer} binary matrix of presence/absence (n x K- dimensional)}

\item{nstar_0}{\link{numeric} initial value of nstar}

\item{p_0}{\link{numeric} initial value of p}

\item{s_0}{\link{numeric} initial value of s}

\item{alpha_bar_0}{\link{numeric} initial value of alpha_bar}

\item{a_star}{\link{numeric}}

\item{b_star}{\link{numeric}}

\item{alpha_p}{\link{numeric}}

\item{beta_p}{\link{numeric}}

\item{a_alpha}{\link{numeric}}

\item{b_alpha}{\link{numeric}}

\item{a_s}{\link{numeric}}

\item{b_s}{\link{numeric}}

\item{tau}{\link{numeric} MALA step-size}

\item{fixed}{\link{logical} vector indicating if the variable is fixed or has prior:
if fixed, it stays equal to initial value}

\item{S}{\link{integer} number of iterations for the MCMC algorithm}

\item{n_burnin}{\link{integer} number of iterations for the burn-in}

\item{thin}{\link{integer} thinning}

\item{seed}{\link{integer} seed}
}
\description{
Metropolis-within-Gibbs sampler for BB + NegBin (Gamma prior on nstar)
}
