% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines_negbin_BB_gibbs.R
\name{gibbs_sampler_negbin_geometric_fixed_pars}
\alias{gibbs_sampler_negbin_geometric_fixed_pars}
\title{Metropolis-within-Gibbs sampler for BB + NegBin (Geometric prior on nstar)}
\usage{
gibbs_sampler_negbin_geometric_fixed_pars(
  Z,
  s_0,
  alpha_bar_0,
  nstar,
  p,
  a_alpha,
  b_alpha,
  a_s,
  b_s,
  tau,
  S,
  n_burnin,
  thin,
  seed
)
}
\arguments{
\item{Z}{\link{integer} binary matrix of presence/absence (n x K- dimensional)}

\item{s_0}{\link{numeric} initial value of s}

\item{alpha_bar_0}{\link{numeric} initial value of alpha_bar}

\item{a_alpha}{\link{numeric}}

\item{b_alpha}{\link{numeric}}

\item{a_s}{\link{numeric}}

\item{b_s}{\link{numeric}}

\item{tau}{\link{numeric} MALA step-size}

\item{S}{\link{integer} number of iterations for the MCMC algorithm}

\item{n_burnin}{\link{integer} number of iterations for the burn-in}

\item{thin}{\link{integer} thinning}

\item{seed}{\link{integer} seed}

\item{nstar_0}{\link{numeric} initial value of nstar}

\item{p_0}{\link{numeric} initial value of p}

\item{q_star}{\link{numeric} hyperparameter of Geometric prior on nstar}

\item{alpha_p}{\link{numeric}}

\item{beta_p}{\link{numeric}}

\item{fixed}{\link{logical} vector indicating if the variable is fixed or has prior:
if fixed, it stays equal to initial value}
}
\description{
Metropolis-within-Gibbs sampler for BB + NegBin (Geometric prior on nstar)
}
