% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines_poiss_BB_gibbs.R
\name{gibbs_sampler_poiss}
\alias{gibbs_sampler_poiss}
\title{Metropolis-within-Gibbs sampler for BB + Poiss}
\usage{
gibbs_sampler_poiss(
  Z,
  lambda_0,
  alpha_bar_0,
  s_0,
  a_l,
  b_l,
  a_alpha,
  b_alpha,
  a_s,
  b_s,
  tau,
  S,
  n_burnin,
  thin,
  seed
)
}
\arguments{
\item{Z}{\link{integer} binary matrix of presence/absence (n x K- dimensional)}

\item{lambda_0}{\link{numeric} initial value of lambda}

\item{alpha_bar_0}{\link{numeric} initial value of alpha_bar}

\item{s_0}{\link{numeric} initial value of s}

\item{a_l}{\link{numeric}}

\item{b_l}{\link{numeric}}

\item{a_alpha}{\link{numeric}}

\item{b_alpha}{\link{numeric}}

\item{a_s}{\link{numeric}}

\item{b_s}{\link{numeric}}

\item{tau}{\link{numeric} MALA step-size}

\item{S}{\link{integer} number of iterations for the MCMC algorithm}

\item{n_burnin}{\link{integer} number of iterations for the burn-in}

\item{thin}{\link{integer} thinning}

\item{seed}{\link{integer} seed}
}
\description{
Metropolis-within-Gibbs sampler for BB + Poiss
}
