% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{improved_chao2_estimator}
\alias{improved_chao2_estimator}
\title{Improved Chao2 estimator}
\usage{
improved_chao2_estimator(data_mat)
}
\arguments{
\item{obs_n}{\link{list} number of new observed}

\item{est_n}{\link{list} number of estimated new}

\item{obs_t}{\link{numeric} number of observed in training set}
}
\description{
Improved Chao2 estimator
}
