% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K_n_r.R
\name{K_n_r.array}
\alias{K_n_r.array}
\title{Sample-based K_n_r curves for a vector of species abundances
(if n_reordering=1, then it is just the observed curves, with no reshuffling)}
\usage{
\method{K_n_r}{array}(object, n_reorderings = 1, seed = 1234)
}
\arguments{
\item{object}{The \code{n x K}-dimensional binary matrix}

\item{n_reorderings}{number of reorderings of the data to average over}

\item{seed}{}
}
\description{
Sample-based K_n_r curves for a vector of species abundances
(if n_reordering=1, then it is just the observed curves, with no reshuffling)
}
