% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K_n_r.R
\name{K_n_r.GibbsFA}
\alias{K_n_r.GibbsFA}
\title{Model-based K_n_r for the GibbsFA models}
\usage{
\method{K_n_r}{GibbsFA}(object, ...)
}
\arguments{
\item{object}{An object of class \code{GibbsFA}}

\item{seed}{Set seed for randomness}
}
\description{
Model-based K_n_r for the GibbsFA models
}
\details{
Perform estimate on K_n_r
}
