% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K_n_r.R
\name{K_n_r}
\alias{K_n_r}
\title{Model-based K_n_r}
\usage{
K_n_r(object, ...)
}
\arguments{
\item{object}{An object of class \code{GibbsFA}}

\item{...}{Additional parameters}
}
\description{
Model-based K_n_r
}
