% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_choice.R
\name{log_posterior_NegBinBB_cpp}
\alias{log_posterior_NegBinBB_cpp}
\title{Log-evaluation of un-normalized NegBinBB posterior}
\usage{
log_posterior_NegBinBB_cpp(post_sample, data_full)
}
\arguments{
\item{post_sample}{contains posterior samples}

\item{data_full}{contains data and all hyperparameters}
}
\description{
Log-evaluation of un-normalized NegBinBB posterior
}
