% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{missed_gt}
\alias{missed_gt}
\title{Good-Toulmin prediction of number of features in extrapolated sample - missed}
\usage{
missed_gt(N, M, sfs, alternative = 0)
}
\arguments{
\item{sfs}{\link{numeric} number of features observed in the training sample}

\item{cts}{\link{array} vector with cumulative number of features in the training sample (from 1 individual to N individual)}
}
\description{
Good-Toulmin prediction of number of features in extrapolated sample - missed
}
