% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines_poiss_BB.R
\name{mm_obj_cv_poiss_BB_rep}
\alias{mm_obj_cv_poiss_BB_rep}
\title{MM objective function with cross-validation for BB with Poisson mixture
with reparametrization}
\usage{
mm_obj_cv_poiss_BB_rep(pars, reord_idx, card_folds, obs_sample)
}
\arguments{
\item{pars}{\link{numeric} pars\link{1} = value of alpha in product-form feature allocation,
pars\link{2} = value of s = theta+alpha in product-form feature allocation,
pars\link{3} =  value of lambda - Poisson hyperparameter}

\item{reord_idx}{\link{numeric} order of the observations to be considered}

\item{card_folds}{\link{numeric} cardinalities of the folds}

\item{obs_sample}{\link{list} list with the whole sample of features, where
$features contains the simulated features for each customer,
$num_new contains the number of new features selected for each customer
$counts contains the counts for the observed features#'}
}
\value{
value of the MM objective function with cross-validation to be
minimized
}
\description{
MM objective function with cross-validation for BB with Poisson mixture
with reparametrization
}
