% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines_negbin_BB.R
\name{mm_obj_negbin_mv_BB_rep}
\alias{mm_obj_negbin_mv_BB_rep}
\title{MM objective function for BB with NB mixture with reparametrization}
\usage{
mm_obj_negbin_mv_BB_rep(pars, ntest, ntrain, val_rep, data_list, seed_id)
}
\arguments{
\item{pars}{\link{numeric} pars\link{1} = value of alpha in product-form feature allocation,
pars\link{2} = value of s = theta+alpha in product-form feature allocation,
pars\link{3} =  value of mu - NB hyperparameter,
pars\link{4} = value of t = sigma2 - mu}

\item{ntest}{\link{integer} dimension of the validation set}

\item{ntrain}{\link{integer} dimension of the training set}

\item{val_rep}{\link{integer} number of samples in the validation averaged over for
assessing the number of new features observed}

\item{data_list}{\link{list} list of features by individual}

\item{seed_id}{\link{numeric} Seed to be used}
}
\value{
value of the MM objective function to be minimized
}
\description{
MM objective function for BB with NB mixture with reparametrization
}
