% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines_poiss_BB.R
\name{mm_obj_poiss_BB_rep}
\alias{mm_obj_poiss_BB_rep}
\title{MM objective function for BB with Poisson mixture with reparametrization}
\usage{
mm_obj_poiss_BB_rep(pars, ntest, ntrain, num_new_m)
}
\arguments{
\item{pars}{\link{numeric} pars\link{1} = value of alpha in product-form feature allocation,
pars\link{2} = value of s = theta+alpha in product-form feature allocation,
pars\link{3} =  value of lambda - Poisson hyperparameter}

\item{ntrain}{\link{integer} dimension of the training set}

\item{num_new_m}{\link{numeric} vector of number of new features selected
by each customer, for the last m customers}

\item{n}{\link{integer} dimension of the observed sample}
}
\value{
value of the MM objective function to be minimized
}
\description{
MM objective function for BB with Poisson mixture with reparametrization
}
