% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{neg_log_EFPF_negbin_BB_rep}
\alias{neg_log_EFPF_negbin_BB_rep}
\title{Negative Log-EFPF for BB with Negative-Binomial mixture with reparametrization}
\usage{
neg_log_EFPF_negbin_BB_rep(n, counts, pars)
}
\arguments{
\item{n}{dimension of the observed sample}

\item{counts}{vector of cardinalities for the observed features}

\item{pars}{pars\link{0} = value of alpha in product-form feature allocation,
pars\link{1} = value of s = theta+alpha in product-form feature allocation,
pars\link{2} =  value of nstar - NegBin hyperparameter,
pars\link{3} = value of p - NegBin hyperparameter}
}
\value{
value of the negative logarithm of the EFPF for the sample of
dimensionality n described by counts
}
\description{
Negative Log-EFPF for BB with Negative-Binomial mixture with reparametrization
}
