% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{neg_log_EFPF_negbin_mv_BB}
\alias{neg_log_EFPF_negbin_mv_BB}
\title{Negative Log-EFPF for BB with Negative-Binomial mixture NO reparametrization -
\itemize{
\item (mean-variance parametrization)
}}
\usage{
neg_log_EFPF_negbin_mv_BB(pars, n, counts, nopt_par, opt)
}
\arguments{
\item{pars}{vector of parameters to optimize}

\item{n}{dimension of the observed sample}

\item{counts}{vector of cardinalities for the observed features}

\item{nopt_par}{vector of parameters not to optimize}

\item{opt}{boolean vector indicating if the variable is optimized}
}
\value{
value of the negative logarithm of the EFPF for the sample of
dimensionality n described by counts
}
\description{
Negative Log-EFPF for BB with Negative-Binomial mixture NO reparametrization -
\itemize{
\item (mean-variance parametrization)
}
}
