% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{neg_log_EFPF_negbin_mv_BB_rep_all}
\alias{neg_log_EFPF_negbin_mv_BB_rep_all}
\title{(when all the parameters are optimized -  mean-variance parametrization)}
\usage{
neg_log_EFPF_negbin_mv_BB_rep_all(pars, n, counts)
}
\arguments{
\item{pars}{vector of parameters to optimize}

\item{n}{dimension of the observed sample}

\item{counts}{vector of cardinalities for the observed features}
}
\value{
value of the negative logarithm of the EFPF for the sample of
dimensionality n described by counts
}
\description{
(when all the parameters are optimized -  mean-variance parametrization)
}
