% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{neg_log_EFPF_NegBinBB}
\alias{neg_log_EFPF_NegBinBB}
\title{Negative Log-EFPF for BB with NB(n0,p) mixture}
\usage{
neg_log_EFPF_NegBinBB(n, counts, pars)
}
\arguments{
\item{n}{dimension of the observed sample}

\item{counts}{vector of cardinalities for the observed features}

\item{pars}{pars\link{0} = value of alpha in product-form feature allocation,
pars\link{1} = value of s = theta + alpha in product-form feature allocation,
pars\link{2} = value of n0 - NB hyperparameter, first
pars\link{3} = value of p - NB hyperparameter, second}
}
\value{
value of the negative logarithm of the EFPF for the sample of
dimensionality n described by counts
}
\description{
Negative Log-EFPF for BB with NB(n0,p) mixture
}
