% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines_gamma_IBP_gibbs.R
\name{neg_log_posterior_gamma_ibp}
\alias{neg_log_posterior_gamma_ibp}
\title{Function for the log posterior density of alpha_hat and s_hat}
\usage{
neg_log_posterior_gamma_ibp(
  pars,
  n,
  K,
  counts,
  a,
  b,
  a_s,
  b_s,
  a_alpha,
  b_alpha
)
}
\arguments{
\item{n}{}

\item{K}{}

\item{counts}{}

\item{a_s}{}

\item{b_s}{}

\item{a_alpha}{}

\item{b_alpha}{}

\item{s_hat}{}

\item{alpha_bar_hat}{}

\item{lambda}{}
}
\description{
Function for the log posterior density of alpha_hat and s_hat
}
