% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines_poiss_BB_gibbs.R
\name{neg_log_posterior_poiss}
\alias{neg_log_posterior_poiss}
\title{Function for the log posterior density of alpha_hat and s_hat}
\usage{
neg_log_posterior_poiss(pars, n, K, counts, lambda, a_s, b_s, a_alpha, b_alpha)
}
\arguments{
\item{n}{}

\item{K}{}

\item{counts}{}

\item{lambda}{}

\item{a_s}{}

\item{b_s}{}

\item{a_alpha}{}

\item{b_alpha}{}

\item{s_hat}{}

\item{alpha_bar_hat}{}
}
\description{
Function for the log posterior density of alpha_hat and s_hat
}
