% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{perc_accuracy}
\alias{perc_accuracy}
\title{Percentage accuracy in prediction}
\usage{
perc_accuracy(train_list, test_list, est_new_features)
}
\arguments{
\item{train_list}{\link{list} list containing the training observations}

\item{test_list}{\link{list} list containing the test observations}

\item{est_new_features}{\link{numeric} estimated number of new features in the test set
(computed with N=#train_list and M=#test_list)}
}
\value{
the percentage accuracy of the estimate on the test set
}
\description{
This function computes the percentage accuracy of the proposed method on
the test set
}
