% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GibbsModels.R
\name{plot.GibbsFA}
\alias{plot.GibbsFA}
\title{Plot for both the mixtures of Beta-Bernoulli and the Gamma IBP}
\usage{
\method{plot}{GibbsFA}(x, type, bw = 1, n_reorderings = 1, M = NULL, seed = 1234, ...)
}
\arguments{
\item{x}{An object of class \code{GibbsFA}.}

\item{type}{Type of plot. Available options are:  \code{"rarefaction"}, \code{"extrapolation"} and \code{"richness"}.
Type \code{"richness"} is available only for mixtures of Beta-Bernoulli.}

\item{bw}{bandwidth for sample-based plot of richness.}

\item{n_reorderings}{number of reorderings for accumularion curve.}

\item{M}{additional sample to predict. Valid only for \code{type = "extrapolation"}. By default, it is equal to the sample size.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Plot for both the mixtures of Beta-Bernoulli and the Gamma IBP
}
\details{
The method produces summary plots for the GibbsFA output
}
