% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_binary_matrix}
\alias{plot_binary_matrix}
\title{Plot Binary matrix of the feature observed for each individual}
\usage{
plot_binary_matrix(mat, max_f = NULL)
}
\arguments{
\item{mat}{\link{matrix} binary matrix (#individuals x #features)}

\item{max_f}{\link{numeric} optional argument indicating the maximum
number of columns to plot}
}
\description{
This function allows to plot the features observed for each individual
following the order-of-appearance
}
