% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_Kmn_given_sample_with_observed}
\alias{plot_Kmn_given_sample_with_observed}
\title{Plot function for the credible intervals of Kmn, given initial sample, and the observed}
\usage{
plot_Kmn_given_sample_with_observed(N, data_list, ci)
}
\arguments{
\item{N}{\link{integer} dimension of initial sample}

\item{data_list}{\link{list} list of features in the whole sample}

\item{ci}{\link{list} it contains means, upper-bounds and lower-bounds of the credible intervals}
}
\description{
This function allows to plot the credible intervals of Kmn given initial sample,
and the observed test as well
}
