% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_Kmn_given_sample_with_observed_both}
\alias{plot_Kmn_given_sample_with_observed_both}
\title{Plot function for the credible intervals of Kmn, given initial sample, for
both efpf and MM, and the observed trajectory}
\usage{
plot_Kmn_given_sample_with_observed_both(N, data_list, ci_efpf, ci_mm)
}
\arguments{
\item{N}{\link{integer} dimension of initial sample}

\item{data_list}{\link{list} list of features in the whole sample}

\item{ci_efpf}{\link{list} it contains means, upper-bounds and lower-bounds of the CI for efpf}

\item{ci_mm}{\link{list} it contains means, upper-bounds and lower-bounds of the CI for mm}
}
\description{
This function allows to plot the credible intervals of Kmn given initial sample,
for both efpf and MM, and the observed test as well
}
