% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_Kmn_median}
\alias{plot_Kmn_median}
\title{Plot function for the credible intervals of Kmn}
\usage{
plot_Kmn_median(ci)
}
\arguments{
\item{ci}{\link{list} it contains medians, upper-bounds and lower-bounds of the credible intervals}
}
\description{
This function allows to plot the credible intervals of Kmn
}
