% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_Kmn_median_pred_all_sp_noibp}
\alias{plot_Kmn_median_pred_all_sp_noibp}
\title{Plot function for the credible intervals of Kmn, given initial sample, and the observed,
for all the 3 Models}
\usage{
plot_Kmn_median_pred_all_sp_noibp(data_list, ci_poiss, ci_negbin, ci_sp, n_avg)
}
\arguments{
\item{ci_poiss}{\link{list} it contains means, upper-bounds and lower-bounds of the CI for Poisson}

\item{ci_negbin}{\link{list} it contains means, upper-bounds and lower-bounds of the CI for NB}

\item{n_avg}{\link{integer}}

\item{train_list}{\link{list}}

\item{test_list}{\link{list}}

\item{ci_ibp}{\link{list} it contains means, upper-bounds and lower-bounds of the CI for IBP}
}
\description{
This function allows to plot the credible intervals of Kmn given initial sample,
and the observed test as well
}
