% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_Kmn_median_pred_and_test}
\alias{plot_Kmn_median_pred_and_test}
\title{Plot function for the credible intervals of Kmn, given initial sample, and the observed}
\usage{
plot_Kmn_median_pred_and_test(train_list, test_list, ci, n_avg)
}
\arguments{
\item{ci}{\link{list} it contains means, upper-bounds and lower-bounds of the credible intervals}

\item{N}{\link{integer} dimension of initial sample}

\item{data_list}{\link{list} list of features in the whole sample}
}
\description{
This function allows to plot the credible intervals of Kmn given initial sample,
and the observed test as well
}
