% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_Kn_median_and_rarefaction}
\alias{plot_Kn_median_and_rarefaction}
\title{Plot function for the credible intervals of Kmn, given initial sample}
\usage{
plot_Kn_median_and_rarefaction(train_list, ci, n_avg)
}
\arguments{
\item{ci}{\link{list} it contains means, upper-bounds and lower-bounds of the credible intervals}

\item{data_list}{\link{list} list of features in the initial sample}
}
\description{
This function allows to plot the credible intervals of Kmn given initial sample
}
