% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_Kn_median_and_rarefaction_all_sp}
\alias{plot_Kn_median_and_rarefaction_all_sp}
\title{Plot function for the credible intervals IN-SAMPLE of Kn, and the observed,
for all the 3 Models}
\usage{
plot_Kn_median_and_rarefaction_all_sp(
  train_list,
  ci_poiss,
  ci_negbin,
  ci_ibp,
  ci_sp,
  n_avg
)
}
\arguments{
\item{train_list}{\link{list}}

\item{ci_poiss}{\link{list} it contains means, upper-bounds and lower-bounds of the CI for Poisson}

\item{ci_negbin}{\link{list} it contains means, upper-bounds and lower-bounds of the CI for NB}

\item{ci_ibp}{\link{list} it contains means, upper-bounds and lower-bounds of the CI for IBP}

\item{n_avg}{\link{integer}}
}
\description{
This function allows to plot the credible intervals of Kn, a posteriori,
and the observed data as well
}
