% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_Kn_median_and_sample}
\alias{plot_Kn_median_and_sample}
\title{Plot function for the credible intervals of Kmn, given initial sample}
\usage{
plot_Kn_median_and_sample(data_list, ci)
}
\arguments{
\item{data_list}{\link{list} list of features in the initial sample}

\item{ci}{\link{list} it contains means, upper-bounds and lower-bounds of the credible intervals}
}
\description{
This function allows to plot the credible intervals of Kmn given initial sample
}
