% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_trajectory}
\alias{plot_trajectory}
\title{Plot function for the observed trajectory}
\usage{
plot_trajectory(data_list)
}
\arguments{
\item{data_list}{\link{list} list of features in the whole sample}
}
\description{
This function allows to plot the observed trajectory
}
