% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{predict_good_toulmin}
\alias{predict_good_toulmin}
\title{Smoothed Good-Toulmin (Chakraborty 2019)
prediction of number of features in extrapolated sample}
\usage{
predict_good_toulmin(N, M, sfs, cts, alternative = 0)
}
\arguments{
\item{sfs}{\link{numeric} number of features observed in the training sample}

\item{cts}{\link{array} vector with cumulative number of features in the training sample (from 1 individual to N individual)}
}
\description{
Smoothed Good-Toulmin (Chakraborty 2019)
prediction of number of features in extrapolated sample
}
