% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GibbsModels_miscellanea.R
\name{prediction.GammaIBP}
\alias{prediction.GammaIBP}
\title{Model-based prediction for IBP with Gamma(a,b) mixture,
with prior on a and b}
\usage{
\method{prediction}{GammaIBP}(object, n = 0, Kn = 0, M = 10, only_last = FALSE, seed = 1234)
}
\arguments{
\item{object}{An object of class \code{GibbsFA, GammaIBP}}

\item{n}{Size of the hypothetical observed sample}

\item{Kn}{Number of distinct features hypothetically observed in the sample}

\item{M}{Maximum size of future sample to extrapolate}

\item{only_last}{logical; TRUE if only extrapolation for the additional sample of size M}

\item{seed}{Set seed for randomness}
}
\description{
Model-based prediction for IBP with Gamma(a,b) mixture,
with prior on a and b
}
\details{
Draw samples from the posterior distribution Kn + K^(n)_m, for different m,
for IBP with Gamma(a,b) mixture, with prior on a and b
}
