% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GibbsModels_miscellanea.R
\name{prediction.GammaIBP_eb}
\alias{prediction.GammaIBP_eb}
\title{Model-based prediction for IBP with Gamma(a,b) mixture - EB version}
\usage{
\method{prediction}{GammaIBP_eb}(object, n = 0, Kn = 0, M = 10, only_last = FALSE)
}
\arguments{
\item{object}{An object of class \code{GibbsFA, GammaIBP}}

\item{n}{Size of the hypothetical observed sample}

\item{Kn}{Number of distinct features hypothetically observed in the sample}

\item{M}{Maximum size of future sample to extrapolate}

\item{only_last}{logical; TRUE if only extrapolation for the additional sample of size M}
}
\description{
Model-based prediction for IBP with Gamma(a,b) mixture - EB version
}
\details{
Return the (n0,mu0) sequences of the NB posterior distribution K^(n)_m,
for different m, for IBP with Gamma(a,b) mixture
}
