% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GibbsModels_miscellanea.R
\name{prediction.PoissonBB}
\alias{prediction.PoissonBB}
\title{Model-based prediction for BB with Poisson(lambda) mixture}
\usage{
\method{prediction}{PoissonBB}(object, n = 0, Kn = 0, M = 10, only_last = FALSE, seed = 1234)
}
\arguments{
\item{object}{An object of class \code{GibbsFA, PoissonBB}}

\item{n}{Size of the hypothetical observed sample}

\item{Kn}{Number of distinct features hypothetically observed in the sample}

\item{M}{Maximum size of future sample to extrapolate}

\item{only_last}{logical; TRUE if only extrapolation for the additional sample of size M}

\item{seed}{Set seed for randomness}
}
\description{
Model-based prediction for BB with Poisson(lambda) mixture
}
\details{
Draw samples from the posterior distribution Kn + K^(n)_m, for different m,
for BB with Poisson(lambda) mixture
}
