% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GibbsModels_miscellanea.R
\name{prediction.PoissonBB_eb}
\alias{prediction.PoissonBB_eb}
\title{Model-based prediction for BB with Poisson(lambda) mixture - EB version}
\usage{
\method{prediction}{PoissonBB_eb}(object, n = 0, Kn = 0, M = 10, only_last = FALSE)
}
\arguments{
\item{object}{An object of class \code{GibbsFA, PoissonBB}}

\item{n}{Size of the hypothetical observed sample}

\item{Kn}{Number of distinct features hypothetically observed in the sample}

\item{M}{Maximum size of future sample to extrapolate}

\item{only_last}{logical; TRUE if only extrapolation for the additional sample of size M}
}
\description{
Model-based prediction for BB with Poisson(lambda) mixture - EB version
}
\details{
Return lambdas sequence of the posterior distribution K^(n)_m, for different m,
for BB with Poisson(lambda) mixture
}
