% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GibbsModels_miscellanea.R
\name{prediction}
\alias{prediction}
\title{Model-based extrapolation}
\usage{
prediction(object, ...)
}
\arguments{
\item{object}{An object of class \code{GibbsFA}}

\item{...}{Additional parameters}
}
\description{
Model-based extrapolation
}
