% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K_n_r.R
\name{prediction_K_n_r.GammaIBP_eb}
\alias{prediction_K_n_r.GammaIBP_eb}
\title{Model-based K_n_r for IBP with Gamma(a,b) mixture - EB version}
\usage{
\method{prediction_K_n_r}{GammaIBP_eb}(object, n)
}
\arguments{
\item{object}{An object of class \code{GibbsFA, GammaIBP}}

\item{n}{Size of the hypothetical observed sample}
}
\description{
Model-based K_n_r for IBP with Gamma(a,b) mixture - EB version
}
\details{
Return list of the (n0,mu0) sequences of the NB for K_n_r,
for different n, for IBP with Gamma(a,b) mixture
}
