% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K_n_r.R
\name{prediction_K_n_r}
\alias{prediction_K_n_r}
\title{Model-based extrapolation}
\usage{
prediction_K_n_r(object, ...)
}
\arguments{
\item{object}{An object of class \code{GibbsFA}}

\item{...}{Additional parameters}
}
\description{
Model-based extrapolation
}
