% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefaction.R
\name{rarefaction.array}
\alias{rarefaction.array}
\title{Sample-based rarefaction curve for a vector of species abundances
(if n_reordering=1, then it is just the observed accumulation curve, with no reshuffling)}
\usage{
\method{rarefaction}{array}(object, n_reorderings = 1, seed = 1234)
}
\arguments{
\item{object}{The \code{n x K}-dimensional binary matrix}

\item{n_reorderings}{number of reorderings of the data to average over}

\item{seed}{}
}
\description{
Sample-based rarefaction curve for a vector of species abundances
(if n_reordering=1, then it is just the observed accumulation curve, with no reshuffling)
}
