% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefaction.R
\name{rarefaction.GammaIBP}
\alias{rarefaction.GammaIBP}
\title{Model-based rarefaction for IBP with Gamma(a,b) mixture,
with prior on a and b}
\usage{
\method{rarefaction}{GammaIBP}(object, seed = 1234)
}
\arguments{
\item{object}{An object of class \code{GibbsFA, GammaIBP}}

\item{seed}{Set seed for randomness}
}
\description{
Model-based rarefaction for IBP with Gamma(a,b) mixture,
with prior on a and b
}
\details{
Draw samples from the posterior distribution of K_n,
for IBP with Gamma(a,b) mixture, and prior on a and b
}
