% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefaction.R
\name{rarefaction.GibbsFA}
\alias{rarefaction.GibbsFA}
\title{Model-based rarefaction for the GibbsFA models}
\usage{
\method{rarefaction}{GibbsFA}(object, ...)
}
\arguments{
\item{object}{An object of class \code{GibbsFA}}

\item{seed}{Set seed for randomness}
}
\description{
Model-based rarefaction for the GibbsFA models
}
\details{
Draw samples from the posterior distribution of K_n,
for BB with NB(n0,mu0) mixture
}
