% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefaction.R
\name{rarefaction.NegBinBB}
\alias{rarefaction.NegBinBB}
\title{Model-based rarefaction for BB with NB(n0,mu0) mixture}
\usage{
\method{rarefaction}{NegBinBB}(object, seed = 1234)
}
\arguments{
\item{object}{An object of class \code{GibbsFA, NegBinBB}}

\item{seed}{Set seed for randomness}
}
\description{
Model-based rarefaction for BB with NB(n0,mu0) mixture
}
\details{
Draw samples from the posterior distribution of K_n,
for BB with NB(n0,mu0) mixture
}
