% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefaction.R
\name{rarefaction.PoissonBB}
\alias{rarefaction.PoissonBB}
\title{Model-based rarefaction for BB with Poisson(lambda) mixture}
\usage{
\method{rarefaction}{PoissonBB}(object, seed = 1234)
}
\arguments{
\item{object}{An object of class \code{GibbsFA, PoissonBB}}

\item{seed}{Set seed for randomness}
}
\description{
Model-based rarefaction for BB with Poisson(lambda) mixture
}
\details{
Draw samples from the posterior distribution of K_n,
for BB with Poisson(lambda) mixture
}
