% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefaction.R
\name{rarefaction}
\alias{rarefaction}
\title{Model-based rarefaction}
\usage{
rarefaction(object, ...)
}
\arguments{
\item{object}{An object of class \code{GibbsFA}}

\item{...}{Additional parameters}
}
\description{
Model-based rarefaction
}
