% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rGammaIBP}
\alias{rGammaIBP}
\title{Random generation from IBP with Gamma(a,b) mixture}
\usage{
rGammaIBP(alpha, theta, a, b, n)
}
\arguments{
\item{alpha}{value of alpha in product-form feature allocation}

\item{theta}{value of theta in product-form feature allocation}

\item{a}{hyperparameter "a" of Gamma(a,b) for gamma}

\item{b}{hyperparameter "b" of Gamma(a,b) for gamma}

\item{n}{dimension of the sample to simulate}
}
\value{
list: $features contains the simulated features for each customer,
$num_new contains the number of new features selected for each customer
$counts contains the counts for the observed features
}
\description{
Random generation from IBP with Gamma(a,b) mixture
}
