% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rNegBinBB}
\alias{rNegBinBB}
\title{Random generation from BB with NB(n0,mu0) mixture}
\usage{
rNegBinBB(alpha, theta, n0, mu0, n)
}
\arguments{
\item{alpha}{value of alpha in product-form feature allocation}

\item{theta}{value of theta in product-form feature allocation}

\item{n0}{hyperparameter "n0" of NB(n0,mu0) for N}

\item{mu0}{hyperparameter "mu0" of NB(n0,mu0) for N}

\item{n}{dimension of the sample to simulate}
}
\value{
list: $features contains the simulated features for each customer,
$num_new contains the number of new features selected for each customer,
$counts contains the counts for the observed features
}
\description{
Random generation from BB with NB(n0,mu0) mixture
}
