% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rPoissonBB}
\alias{rPoissonBB}
\title{Random generation from BB with Poisson(lambda) mixture}
\usage{
rPoissonBB(alpha, theta, lambda, n)
}
\arguments{
\item{alpha}{value of alpha in product-form feature allocation}

\item{theta}{value of theta in product-form feature allocation}

\item{lambda}{Poisson hyperparameter}

\item{n}{dimension of the sample to generate}
}
\value{
list: $features contains the simulated features for each customer,
$num_new contains the number of new features selected for each customer
$counts contains the counts for the observed features
}
\description{
Random generation from BB with Poisson(lambda) mixture
}
