% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipf_generator.R
\name{rzipf}
\alias{rzipf}
\title{Simulate observations with features from Zipf distribution}
\usage{
rzipf(n, K, xi, seed = 1234)
}
\arguments{
\item{n}{\link{integer} number of observations to extract}

\item{K}{\link{integer} maximum number of observable features}

\item{xi}{\link{numeric} probability factor}
}
\description{
This function draws samples of features/observations from Zipf distribution
}
