% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GibbsModels_internal.R
\name{sampler_classicBB}
\alias{sampler_classicBB}
\title{Metropolis-within-Gibbs sampler for classic BB}
\usage{
sampler_classicBB(
  Z,
  alpha_bar_0,
  s_0,
  a_alpha,
  b_alpha,
  a_s,
  b_s,
  N,
  tau,
  S,
  n_burnin,
  thin,
  seed
)
}
\arguments{
\item{Z}{\link{integer} binary matrix of presence/absence (\code{n x K} - dimensional)}

\item{alpha_bar_0}{\link{numeric} initial value of alpha_bar (with alpha_bar = - alpha)}

\item{s_0}{\link{numeric} initial value of s (with s = alpha + theta)}

\item{a_alpha}{\link{numeric} hyperparameter "a" of Gamma(a,b) for alpha_bar}

\item{b_alpha}{\link{numeric} hyperparameter "b" of Gamma(a,b) for alpha_bar}

\item{a_s}{\link{numeric} hyperparameter "a" of Gamma(a,b) for s}

\item{b_s}{\link{numeric} hyperparameter "b" of Gamma(a,b) for s}

\item{N}{\link{numeric} hyperparameter N for classic BB (total number of features)}

\item{tau}{\link{numeric} MALA step-size}

\item{S}{\link{integer} number of iterations for the MCMC algorithm}

\item{n_burnin}{\link{integer} number of burn-in iterations}

\item{thin}{\link{integer} thinning}

\item{seed}{\link{integer} seed}
}
\description{
Metropolis-within-Gibbs sampler for classic BB
}
