% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GibbsModels_internal.R
\name{sampler_classicIBP}
\alias{sampler_classicIBP}
\title{Metropolis-within-Gibbs sampler for classic IBP}
\usage{
sampler_classicIBP(
  Z,
  alpha_0,
  s_0,
  a_alpha,
  b_alpha,
  a_s,
  b_s,
  gam,
  sigq_alpha,
  sigq_s,
  S,
  n_burnin,
  thin,
  seed
)
}
\arguments{
\item{Z}{\link{integer} binary matrix of presence/absence (\code{n x K} - dimensional)}

\item{alpha_0}{\link{numeric} initial value of alpha}

\item{s_0}{\link{numeric} initial value of s (with s = alpha + theta)}

\item{a_alpha}{\link{numeric} hyperparameter "a" of Beta(a,b) for alpha}

\item{b_alpha}{\link{numeric} hyperparameter "b" of Beta(a,b) for alpha}

\item{a_s}{\link{numeric} hyperparameter "a" of Gamma(a,b) for s}

\item{b_s}{\link{numeric} hyperparameter "b" of Gamma(a,b) for s}

\item{gam}{\link{numeric} total mass of the IBP}

\item{sigq_alpha}{\link{numeric} variance proposal for alpha}

\item{sigq_s}{\link{numeric} variance proposal for s}

\item{S}{\link{integer} number of iterations for the MCMC algorithm}

\item{n_burnin}{\link{integer} number of burn-in iterations}

\item{thin}{\link{integer} thinning}

\item{seed}{\link{integer} seed}
}
\description{
Metropolis-within-Gibbs sampler for classic IBP
}
