% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GibbsModels_internal.R
\name{sampler_GammaIBP}
\alias{sampler_GammaIBP}
\title{Metropolis-within-Gibbs sampler for IBP with Gamma(a,b) mixture,
with prior also on a and b}
\usage{
sampler_GammaIBP(
  Z,
  alpha_0,
  s_0,
  a_0,
  b_0,
  a_alpha,
  b_alpha,
  a_s,
  b_s,
  q,
  r,
  t,
  sigq_alpha,
  sigq_s,
  S,
  n_burnin,
  thin,
  seed
)
}
\arguments{
\item{Z}{\link{integer} binary matrix of presence/absence (n x K- dimensional)}

\item{alpha_0}{\link{numeric} initial value of alpha}

\item{s_0}{\link{numeric} initial value of s}

\item{a_0}{\link{numeric} initial value of a}

\item{b_0}{\link{numeric} initial value of b}

\item{a_alpha}{\link{numeric}}

\item{b_alpha}{\link{numeric}}

\item{a_s}{\link{numeric}}

\item{b_s}{\link{numeric}}

\item{q}{\link{numeric} hyperparameter of Geometric prior on a}

\item{r}{\link{numeric}}

\item{t}{\link{numeric}}

\item{S}{\link{integer} number of iterations for the MCMC algorithm}

\item{n_burnin}{\link{integer} number of iterations for the burn-in}

\item{thin}{\link{integer} thinning}

\item{seed}{\link{integer} seed}

\item{tau}{\link{numeric} MALA step-size}

\item{fixed}{\link{logical} vector indicating if the variable is fixed or has prior:
if fixed, it stays equal to initial value - order: a, b, alpha, s}
}
\description{
Metropolis-within-Gibbs sampler for IBP with Gamma(a,b) mixture,
with prior also on a and b
}
