% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GibbsModels_internal.R
\name{sampler_PoissonBB}
\alias{sampler_PoissonBB}
\title{Metropolis-within-Gibbs sampler for BB with Poisson(lambda) mixture}
\usage{
sampler_PoissonBB(
  Z,
  alpha_bar_0,
  s_0,
  a_alpha,
  b_alpha,
  a_s,
  b_s,
  lambda,
  tau,
  S,
  n_burnin,
  thin,
  seed
)
}
\arguments{
\item{Z}{\link{integer} binary matrix of presence/absence (n x K - dimensional)}

\item{alpha_bar_0}{\link{numeric} initial value of alpha_bar}

\item{s_0}{\link{numeric} initial value of s}

\item{a_alpha}{\link{numeric}}

\item{b_alpha}{\link{numeric}}

\item{a_s}{\link{numeric}}

\item{b_s}{\link{numeric}}

\item{lambda}{\link{numeric}}

\item{tau}{\link{numeric} MALA step-size}

\item{S}{\link{integer} number of iterations for the MCMC algorithm}

\item{n_burnin}{\link{integer} number of iterations for the burn-in}

\item{thin}{\link{integer} thinning}

\item{seed}{\link{integer} seed}
}
\description{
Metropolis-within-Gibbs sampler for BB with Poisson(lambda) mixture
}
