% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/richness.R
\name{total_richness.NegBinBB}
\alias{total_richness.NegBinBB}
\title{Model-based richness posterior distribution for BB with NB(n0,mu0) mixture}
\usage{
\method{total_richness}{NegBinBB}(object, seed = 1234)
}
\arguments{
\item{object}{An object of class \code{GibbsFA, NegBinBB}}

\item{seed}{Set seed for randomness}
}
\description{
Model-based richness posterior distribution for BB with NB(n0,mu0) mixture
}
\details{
Draw samples from the posterior distribution of the total richness
for BB with NB(n0,mu0) mixture
}
