% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/richness.R
\name{total_richness.NegBinBB_eb}
\alias{total_richness.NegBinBB_eb}
\title{Model-based richness posterior distribution for BB with NB(n0,mu0) mixture - EB version}
\usage{
\method{total_richness}{NegBinBB_eb}(object)
}
\arguments{
\item{object}{An object of class \code{GibbsFA, NegBinBB_eb}}
}
\description{
Model-based richness posterior distribution for BB with NB(n0,mu0) mixture - EB version
}
\details{
Return the (n0,mu0) parameter of the NB posterior distribution of N'
for BB with NB(n0,mu0) mixture
}
