% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/richness.R
\name{total_richness.PoissonBB}
\alias{total_richness.PoissonBB}
\title{Model-based richness posterior distribution for BB with Poisson(lambda) mixture}
\usage{
\method{total_richness}{PoissonBB}(object, seed = 1234)
}
\arguments{
\item{object}{An object of class \code{GibbsFA, PoissonBB}}

\item{seed}{Set seed for randomness}
}
\description{
Model-based richness posterior distribution for BB with Poisson(lambda) mixture
}
\details{
Draw samples from the posterior distribution of the total richness
for BB with Poisson(lambda) mixture
}
