% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/richness.R
\name{total_richness.PoissonBB_eb}
\alias{total_richness.PoissonBB_eb}
\title{Model-based richness posterior distribution for BB with Poisson(lambda) mixture - EB version}
\usage{
\method{total_richness}{PoissonBB_eb}(object)
}
\arguments{
\item{object}{An object of class \code{GibbsFA, PoissonBB_eb}}
}
\description{
Model-based richness posterior distribution for BB with Poisson(lambda) mixture - EB version
}
\details{
Return the lambda parameter of the Poisson posterior distribution of N'
for BB with Poisson(lambda) mixture
}
