% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/richness.R
\name{total_richness}
\alias{total_richness}
\title{Model-based richness posterior distribution}
\usage{
total_richness(object, ...)
}
\arguments{
\item{object}{An object of class \code{GibbsFA}}

\item{...}{Additional parameters}
}
\description{
Model-based richness posterior distribution
}
