% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_choice.R
\name{compute_AICs_BICs}
\alias{compute_AICs_BICs}
\title{Compute AIC/BIC: available models PoissonBB, NegBinBB, GammaIBP (Empirical-Bayes approach)}
\usage{
compute_AICs_BICs(eb_model_fit)
}
\arguments{
\item{eb_model_fit}{object of class \code{GibbsFA, PoissonBB_eb}
or \code{GibbsFA, NegBinBB_eb} or \code{GibbsFA, GammaIBP_eb}}
}
\description{
Compute AIC/BIC: available models PoissonBB, NegBinBB, GammaIBP (Empirical-Bayes approach)
}
