% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_choice.R
\name{compute_BF}
\alias{compute_BF}
\title{Compute Bayes Factor (BF) for: NegBinBB vs GammaIBP (Fully-Bayesian approach)}
\usage{
compute_BF(NegBinBB_fit, GammaIBP_fit)
}
\arguments{
\item{NegBinBB_fit}{object of class \code{GibbsFA, NegBinBB}}

\item{GammaIBP_fit}{object of class \code{GibbsFA, GammaIBP}}
}
\description{
Compute Bayes Factor (BF) for: NegBinBB vs GammaIBP (Fully-Bayesian approach)
}
