% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_choice.R
\name{compute_log_marginal_likelihood_bridge}
\alias{compute_log_marginal_likelihood_bridge}
\title{Computes log marginal likelihood via bridge sampling: available models NegBinBB, GammaIBP (Fully-Bayesian approach)}
\usage{
compute_log_marginal_likelihood_bridge(model_fit)
}
\arguments{
\item{model_fit}{object of class \code{GibbsFA, NegBinBB} or \code{GibbsFA, GammaIBP}}
}
\description{
Computes log marginal likelihood via bridge sampling: available models NegBinBB, GammaIBP (Fully-Bayesian approach)
}
