% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines_poiss_BB.R
\name{EB_MM_poiss_BB}
\alias{EB_MM_poiss_BB}
\title{EB based on method of moments - BB with Poisson mixture}
\usage{
EB_MM_poiss_BB(n, ntrain, num_new, pars_0)
}
\arguments{
\item{n}{\link{integer} dimension of the observed sample}

\item{ntrain}{\link{integer} dimension of training set}

\item{num_new}{\link{numeric} vector of number of new features selected by each customer}

\item{pars_0}{\link{numeric} Initialization for (alpha, theta, lambda) to
be optimized}
}
\description{
This function returns the value of (alpha, theta, lambda) obtained
via the method of moments for the given sample - BB with Poisson mixture
}
