% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines_poiss_BB_gibbs.R
\name{generate_Ntilde_chain_poiss}
\alias{generate_Ntilde_chain_poiss}
\title{Title}
\usage{
generate_Ntilde_chain_poiss(
  lambda_chain_poiss,
  alpha_chain_poiss,
  theta_chain_poiss,
  n,
  Kn
)
}
\arguments{
\item{lambda_chain_poiss}{}

\item{alpha_chain_poiss}{}

\item{theta_chain_poiss}{}

\item{n}{}

\item{Kn}{}
}
\description{
Title
}
