% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GibbsModels.R
\name{GibbsFA}
\alias{GibbsFA}
\title{Gibbs-type feature allocation models (GibbsFA)}
\usage{
GibbsFA(
  feature_matrix,
  model,
  hyperparams,
  initialization,
  mcmcparams,
  seed = 1234
)
}
\arguments{
\item{feature_matrix}{A \code{n x K}-dimensional binary matrix of features}

\item{model}{Model to fit. Available models are \code{PoissonBB} (BB with Poisson(lambda) mixture),
\code{NegBinBB} (BB with NB(n0, mu0) mixture), \code{GammaIBP} (IBP with Gamma(a, b) mixture)}

\item{...}{Additional parameters for hyperparameter elicitation and MCMC settins}
}
\value{
An object of class \code{GibbsFA}
}
\description{
Gibbs-type feature allocation models (GibbsFA)
}
