% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefaction.R
\name{rarefaction.array}
\alias{rarefaction.array}
\title{Sample-based rarefaction curve for a vector of species abundances
(RIGHT NOW IT IS JUST THE OBSERVED SAMPLE, WITH NO AVERAGE)}
\usage{
\method{rarefaction}{array}(object)
}
\arguments{
\item{object}{The \code{n x K}-dimensional binary matrix}
}
\description{
Sample-based rarefaction curve for a vector of species abundances
(RIGHT NOW IT IS JUST THE OBSERVED SAMPLE, WITH NO AVERAGE)
}
