% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rNegBinBB}
\alias{rNegBinBB}
\title{Random generation from BB with NB(nstar,p) mixture}
\usage{
rNegBinBB(alpha, theta, nstar, p, n)
}
\arguments{
\item{alpha}{value of alpha in product-form feature allocation}

\item{theta}{value of theta in product-form feature allocation}

\item{nstar}{Negative-Binomial hyperparameter (number of successes)}

\item{p}{Negative-Binomial hyperparameter (success probability)}

\item{n}{dimension of the sample to simulate}
}
\value{
list: $features contains the simulated features for each customer,
$num_new contains the number of new features selected for each customer,
$counts contains the counts for the observed features
}
\description{
Random generation from BB with NB(nstar,p) mixture
}
