% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GibbsModels_internal.R
\name{sampler_GammaIBP}
\alias{sampler_GammaIBP}
\title{Metropolis-within-Gibbs sampler for IBP with Gamma(a,b) mixture,
with prior also on a and b}
\usage{
sampler_GammaIBP(
  Z,
  alpha_0,
  s_0,
  a_0,
  b_0,
  a_alpha,
  b_alpha,
  a_s,
  b_s,
  q,
  r,
  t,
  sigq_alpha,
  sigq_s,
  S,
  n_burnin,
  thin,
  seed
)
}
\arguments{
\item{Z}{\link{integer} binary matrix of presence/absence (\code{n x K} - dimensional)}

\item{alpha_0}{\link{numeric} initial value of alpha}

\item{s_0}{\link{numeric} initial value of s (with s = alpha + theta)}

\item{a_0}{\link{numeric} initial value of a of Gamma(a,b) for gamma}

\item{b_0}{\link{numeric} initial value of b of Gamma(a,b) for gamma}

\item{a_alpha}{\link{numeric} hyperparameter "a" of Beta(a,b) for alpha}

\item{b_alpha}{\link{numeric} hyperparameter "b" of Beta(a,b) for alpha}

\item{a_s}{\link{numeric} hyperparameter "a" of Gamma(a,b) for s}

\item{b_s}{\link{numeric} hyperparameter "b" of Gamma(a,b) for s}

\item{q}{\link{numeric} hyperparameter "q" of Geo(q) for a}

\item{r}{\link{numeric} hyperparameter "r" of Gamma(r,t) for b}

\item{t}{\link{numeric} hyperparameter "t" of Gamma(r,t) for b}

\item{S}{\link{integer} number of iterations for the MCMC algorithm}

\item{n_burnin}{\link{integer} number of burn-in iterations}

\item{thin}{\link{integer} thinning}

\item{seed}{\link{integer} seed}

\item{tau}{\link{numeric} MALA step-size}
}
\description{
Metropolis-within-Gibbs sampler for IBP with Gamma(a,b) mixture,
with prior also on a and b
}
